load & save scene

+++ sphere.cpp
@@ -1,7 +1,7 @@
 #include "sphere.hpp"
 
 
-Sphere::Sphere(const Point3D& p, const Color& c, int reflection, int radius):Shape(p, c, reflection), _radius(radius)
+Sphere::Sphere(const Point3D& p, const Color& c, double reflection, double radius):Shape(p, c, reflection), _radius(radius)
 {
 
 }
